arreglos visuales

+++ app/routes/libros.py
@@ -1,8 +1,11 @@
-from flask import Blueprint, render_template, request, redirect, flash, jsonify, url_for
+from flask import Blueprint, render_template, request, redirect, flash, jsonify, url_for,session
 from werkzeug.utils import secure_filename
-from models.controladordatabase import db, Libros, Prestamo,Autor,Genero,Editorial,Votacion
+from models.controladordatabase import db, Libros, Prestamo,Autor,Genero,Editorial,Votacion,Usuario
 from routes.login import login_required, admin_required
 import os
+import requests
+from io import BytesIO
+from PIL import Image
 from sqlalchemy import or_
 from random import sample
 
@@ -42,51 +45,74 @@ def agregarlibro():
     fecha_emision = request.form['fecha_publicacion']
     stock = request.form['stock']
     precio = float(request.form['precio'])
-    precio_alquiler = request.form.get('precio_alquiler', None)  # Puede ser opcional
-    isbn = request.form.get('isbn') or None  # Si el campo está vacío, almacena None    
+    precio_alquiler = request.form.get('precio_alquiler', None)  
+    isbn = request.form.get('isbn') or None     
     idioma = request.form['idioma']
     edicion = request.form['edicion']
     paginas = int(request.form['paginas'])
     formato = request.form['formato']
-    
-    # Manejo de la imagen
-    imagen = request.files['imagen']
-    if imagen:
+
+    imagen = request.files.get('imagen')
+    imagen_url = request.form.get('imagen_url')
+    img_name = None
+
+    if imagen:  # Si se ha cargado una imagen local
         basepath = os.path.dirname(__file__)  # Ruta base del archivo actual
         filename = secure_filename(imagen.filename)  # Obtener el nombre original
         extension = os.path.splitext(filename)[1]  # Obtener la extensión del archivo
         nuevoNombreFile = stringAleatorio() + extension  # Generar un nombre único para la imagen
 
-        # Guardar la imagen en la carpeta static/img
+        
         upload_path = os.path.join(basepath, '../static/img', nuevoNombreFile) 
         imagen.save(upload_path)  
+        img_name = nuevoNombreFile  
 
-        # Crear una nueva instancia del libro
-        nuevolibro = Libros(
-            titulo=titulo,
-            id_autor=id_autor,  # Relación con la tabla Autores
-            id_genero=id_genero,  # Relación con la tabla Géneros
-            id_editorial=id_editorial,  # Relación con la tabla Editoriales
-            descripcion=descripcion,
-            fecha_emision=fecha_emision,
-            stock=stock,
-            precio=precio,
-            precio_alquiler=precio_alquiler if precio_alquiler else None,  # Campo opcional
-            isbn=isbn,
-            idioma=idioma,
-            edicion=edicion,
-            paginas=paginas,
-            formato=formato,
-            img=nuevoNombreFile  # Almacenar el nombre de la imagen
-        )
-
+    elif imagen_url:  
         try:
-            db.session.add(nuevolibro)
-            db.session.commit()
-            flash('Libro agregado exitosamente', 'success')
+            
+            response = requests.get(imagen_url)
+            if response.status_code == 200:
+                basepath = os.path.dirname(__file__)
+                extension = os.path.splitext(imagen_url)[1]  
+                nuevoNombreFile = stringAleatorio() + extension  
+                
+                
+                upload_path = os.path.join(basepath, '../static/img', nuevoNombreFile)
+                image = Image.open(BytesIO(response.content))  
+                image.save(upload_path)
+                img_name = nuevoNombreFile  
+            else:
+                flash('No se pudo descargar la imagen desde la URL', 'danger')
+                return redirect(url_for('libros.bienvenida'))
         except Exception as e:
-            db.session.rollback()
-            return jsonify({'error': str(e)}), 400
+            flash(f'Error al descargar la imagen: {str(e)}', 'danger')
+            return redirect(url_for('libros.bienvenida'))
+
+    nuevolibro = Libros(
+        titulo=titulo,
+        id_autor=id_autor,  # Relación con la tabla Autores
+        id_genero=id_genero, 
+        id_editorial=id_editorial,  
+        descripcion=descripcion,
+        fecha_emision=fecha_emision,
+        stock=stock,
+        precio=precio,
+        precio_alquiler=precio_alquiler if precio_alquiler else None,  
+        isbn=isbn,
+        idioma=idioma,
+        edicion=edicion,
+        paginas=paginas,
+        formato=formato,
+        img=img_name if img_name else "sinimagen.jpg"  #
+    )
+
+    try:
+        db.session.add(nuevolibro)
+        db.session.commit()
+        flash('Libro agregado exitosamente', 'success')
+    except Exception as e:
+        db.session.rollback()
+        flash(f'Error al agregar el libro: {str(e)}', 'danger')
 
     return redirect(url_for('libros.bienvenida'))
 
@@ -142,7 +168,24 @@ def agregar_editorial():
     
     return render_template('libros/agregar_editorial.html')
 
+@libros.route('/eliminareditorial/<int:id>')
+@admin_required
+def eliminareditorial(id):
+    nombre_editorial = request.form['nombre_editorial']
+    nueva_editorial = Editorial(nombre=nombre_editorial)
+    if nueva_editorial:
+        prestamo_activo = Prestamo.query.filter_by(id_libro=id).first()
 
+        if prestamo_activo:
+            flash(f'El libro  no puede ser eliminado porque está prestado.', 'error')
+        else:
+            db.session.delete(prestamo_activo)
+            db.session.commit()
+            flash(f'Libro  eliminado con éxito', 'success')
+    else:
+        flash('Libro no encontrado', 'error')
+
+    return redirect(url_for('libros.verlibro'))
 
 
 @libros.route('/eliminarlibro/<int:id>')
@@ -174,18 +217,21 @@ def editarlibro(id):
 
     if request.method == "POST":
         libro.titulo = request.form['titulo']
-        libro.autor_id = int(request.form['autor'])  # Actualizar la relación con el autor
-        libro.genero_id = int(request.form['genero'])  # Actualizar la relación con el género
-        libro.editorial_id = int(request.form['editorial']) if request.form['editorial'] else None
+        libro.id_autor = int(request.form['autor'])  
+        libro.id_genero = int(request.form['genero'])  
+        libro.id_editorial = int(request.form['editorial']) if request.form['editorial'] else None
         libro.descripcion = request.form['descripcion']
         libro.fecha_emision = request.form['fecha_publicacion']
         libro.stock = int(request.form['stock'])
         libro.precio = float(request.form['precio'])
         libro.precio_alquiler = float(request.form['precio_alquiler']) if request.form['precio_alquiler'] else None
 
+        imagen_url = request.form.get('imagen_url') 
         nueva_imagen = request.files.get('imagen')
+        basepath = os.path.dirname(__file__)
+
+        # Comprobar si se subió una nueva imagen o se proporcionó una URL
         if nueva_imagen:
-            basepath = os.path.dirname(__file__)
             filename = secure_filename(nueva_imagen.filename)
             extension = os.path.splitext(filename)[1]
             nuevoNombreFile = stringAleatorio() + extension
@@ -193,14 +239,37 @@ def editarlibro(id):
             upload_path = os.path.join(basepath, '../static/img', nuevoNombreFile)
             nueva_imagen.save(upload_path)
 
-            # Opción para eliminar la imagen anterior (si existe)
-            if libro.img:
+            # Si el libro ya tiene una imagen, eliminarla
+            if libro.img and libro.img != "sinimagen.jpg":
                 imagen_antigua = os.path.join(basepath, '../static/img', libro.img)
                 if os.path.exists(imagen_antigua):
                     os.remove(imagen_antigua)
 
-            # Actualizar el nombre de la imagen en la base de datos
             libro.img = nuevoNombreFile
+
+        elif imagen_url:
+            try:
+                response = requests.get(imagen_url)
+                response.raise_for_status()  # Verificar si la URL es válida
+                img_bytes = BytesIO(response.content)
+                img = Image.open(img_bytes)
+
+                extension = os.path.splitext(imagen_url)[1]
+                nuevoNombreFile = stringAleatorio() + extension
+                upload_path = os.path.join(basepath, '../static/img', nuevoNombreFile)
+                img.save(upload_path)
+
+                # Si el libro ya tiene una imagen, eliminarla
+                if libro.img and libro.img != "sinimagen.jpg":
+                    imagen_antigua = os.path.join(basepath, '../static/img', libro.img)
+                    if os.path.exists(imagen_antigua):
+                        os.remove(imagen_antigua)
+
+                libro.img = nuevoNombreFile
+
+            except Exception as e:
+                flash(f"Error al descargar la imagen: {str(e)}", "danger")
+                return redirect(url_for('libros.bienvenida'))
 
         try:
             db.session.commit()
@@ -225,10 +294,8 @@ def verlibro():
     editorial = request.args.get('editorial', '', type=str)
     genero = request.args.get('genero', '', type=str)
 
-   # Consulta básica de libros
     query = Libros.query
 
-    # Aplicar filtros de búsqueda
     if titulo:
         query = query.filter(Libros.titulo.ilike(f'%{titulo}%'))
     if autor:
@@ -241,30 +308,27 @@ def verlibro():
     pagination = query.paginate(page=page, per_page=per_page)
 
 
-    # Obtener todos los autores, editoriales y géneros para los menús desplegables
-    autores = Autor.query.all()
-    editoriales = Editorial.query.all()
-    generos = Genero.query.all()
+    autores = Autor.query.order_by(Autor.nombre.asc(), Autor.apellido.asc()).all()
+
+    generos = Genero.query.order_by(Genero.nombre.asc()).all()
+
+    editoriales = Editorial.query.order_by(Editorial.nombre.asc()).all()
 
     return render_template('libros/verlibros.html', pagination=pagination, autores=autores, editoriales=editoriales, generos=generos)
 
 
 @libros.route('/libros', methods=['GET'])
 def libross():
-    # Obtener los parámetros de filtro desde la URL
     titulo = request.args.get('titulo', '', type=str)
     autor = request.args.get('autor', '', type=str)
     editorial = request.args.get('editorial', '', type=str)
     genero = request.args.get('genero', '', type=str)
 
-    # Paginación
     page = request.args.get('page', 1, type=int)
-    per_page = 10  # Número de libros por página
+    per_page = 10  
 
-    # Consulta básica de libros
     query = Libros.query
 
-    # Aplicar filtros de búsqueda
     if titulo:
         query = query.filter(Libros.titulo.ilike(f'%{titulo}%'))
     if autor:
@@ -274,13 +338,14 @@ def libross():
     if genero:
         query = query.join(Genero).filter(Genero.nombre.ilike(f'%{genero}%'))
 
-    # Aplicar la paginación
     pagination = query.paginate(page=page, per_page=per_page)
 
-    # Obtener todos los autores, editoriales y géneros para los menús desplegables
-    autores = Autor.query.all()
-    editoriales = Editorial.query.all()
-    generos = Genero.query.all()
+       
+    autores = Autor.query.order_by(Autor.nombre.asc(), Autor.apellido.asc()).all()
+    
+    generos = Genero.query.order_by(Genero.nombre.asc()).all()
+    
+    editoriales = Editorial.query.order_by(Editorial.nombre.asc()).all()
 
     return render_template('libros/libros.html', pagination=pagination, autores=autores, editoriales=editoriales, generos=generos)
 
@@ -288,9 +353,18 @@ def libross():
 def detalle(libro_id):
     libro = Libros.query.get_or_404(libro_id)
     promedio_calificacion = db.session.query(db.func.avg(Votacion.calificacion)).filter(Votacion.id_libro == libro_id).scalar()
-    comentarios = libro.comentarios  # Recupera los comentarios asociados a este libro
+    
+    if "nombre" not in session:
+            return render_template('libros/detalle.html', libro=libro,promedio_calificacion=promedio_calificacion)
 
-    return render_template('libros/detalle.html', libro=libro,promedio_calificacion=promedio_calificacion,comentarios=comentarios)
+    else:
+        comentarios = libro.comentarios  
+        usuario_id = Usuario.query.filter_by(nombre=session['nombre']).first().id
+
+    votacion = Votacion.query.filter_by(id_libro=libro_id, id_usuario=usuario_id).first()
+
+
+    return render_template('libros/detalle.html', libro=libro,promedio_calificacion=promedio_calificacion,comentarios=comentarios,votacion=votacion)
 
 
 @libros.route("/terminosycondiciones")
